clazzy_v2/requirements.txt
# Clazzy V2 - Requirements
# Production-grade dependencies for the intelligent fashion assistant

# =====================
# Core Framework
# =====================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# =====================
# Machine Learning
# =====================
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0        # For CLIP
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# =====================
# Database & Storage
# =====================
sqlalchemy>=2.0.0
alembic>=1.13.0             # Migrations
redis>=5.0.0                # Optional caching

# =====================
# External Services
# =====================
httpx>=0.26.0               # Async HTTP client
anthropic>=0.18.0           # Claude API (optional)
openai>=1.10.0              # OpenAI API (optional)

# =====================
# Image Processing
# =====================
opencv-python>=4.8.0
colorthief>=0.2.1           # Color extraction alternative

# =====================
# Vector Store (Optional)
# =====================
# faiss-cpu>=1.7.4          # For similarity search
# chromadb>=0.4.0           # Alternative vector DB

# =====================
# Development
# =====================
python-multipart>=0.0.6     # File uploads
python-dotenv>=1.0.0        # Environment variables

# =====================
# Testing (dev only)
# =====================
# pytest>=7.4.0
# pytest-asyncio>=0.23.0
# httpx>=0.26.0

# =====================
# Monitoring (optional)
# =====================
# prometheus-client>=0.19.0
# opentelemetry-api>=1.22.0
